requirements.txt
Django==4.0
django_ckeditor==6.5.1
django_imagekit==4.1.0
python-dotenv==1.0.0
